+++ apps/docs-snippets/src/guide/provider/querying-the-chain.test.ts
@@ -10,7 +10,7 @@ describe('querying the chain', () => {
     // #import { Provider, FUEL_NETWORK_URL, generateTestWallet };
     const provider = await Provider.create(FUEL_NETWORK_URL);
     const assetIdA = '0x0101010101010101010101010101010101010101010101010101010101010101';
-    const baseAssetId = await provider.getBaseAssetId();
+    const baseAssetId = provider.getBaseAssetId();
 
     const wallet = await generateTestWallet(provider, [
       [42, baseAssetId],
@@ -54,7 +54,7 @@ describe('querying the chain', () => {
 
     const provider = await Provider.create(FUEL_NETWORK_URL);
     const assetIdA = '0x0101010101010101010101010101010101010101010101010101010101010101';
-    const baseAssetId = await provider.getBaseAssetId();
+    const baseAssetId = provider.getBaseAssetId();
 
     const wallet = await generateTestWallet(provider, [
       [42, baseAssetId],
@@ -79,7 +79,7 @@ describe('querying the chain', () => {
 
     const provider = await Provider.create(FUEL_NETWORK_URL);
     const assetIdA = '0x0101010101010101010101010101010101010101010101010101010101010101';
-    const baseAssetId = await provider.getBaseAssetId();
+    const baseAssetId = provider.getBaseAssetId();
 
     const wallet = await generateTestWallet(provider, [
       [42, baseAssetId],
@@ -110,5 +110,19 @@ describe('querying the chain', () => {
     });
     // #endregion Provider-get-blocks
     expect(blocks.length).toBe(10);
+  });
+
+  it('can getMessageByNonce', async () => {
+    // #region getMessageByNonce
+    // #import { FUEL_NETWORK_URL, Provider };
+
+    const provider = await Provider.create(FUEL_NETWORK_URL);
+
+    const nonce = '0x381de90750098776c71544527fd253412908dec3d07ce9a7367bd1ba975908a0';
+    const message = await provider.getMessageByNonce(nonce);
+
+    expect(message).toBeDefined();
+    expect(message?.nonce).toEqual(nonce);
+    // #endregion getMessageByNonce
   });
 });