+++ packages/fuel-gauge/src/policies.test.ts
@@ -1,7 +1,8 @@
 import { generateTestWallet } from '@fuel-ts/account/test-utils';
-import type { BigNumberish, Transaction, WalletUnlocked } from 'fuels';
+import type { BaseTransactionRequest, BigNumberish, Transaction, WalletUnlocked } from 'fuels';
 import {
   ContractFactory,
+  CreateTransactionRequest,
   FUEL_NETWORK_URL,
   PolicyType,
   Provider,
@@ -58,6 +59,68 @@ describe('Policies', () => {
     expect(bn(transaction.policies?.[3]?.data).lte(bn(params.maxFee))).toBeTruthy();
   };
 
+  it('should ensure optional TX policies are not set when not informed', () => {
+    let txRequest: BaseTransactionRequest = new ScriptTransactionRequest();
+
+    txRequest = new ScriptTransactionRequest();
+
+    expect(txRequest.tip).toBeUndefined();
+    expect(txRequest.maturity).toBeUndefined();
+    expect(txRequest?.witnessLimit).toBeUndefined();
+
+    expect(txRequest.maxFee).toBeDefined();
+
+    let tx = txRequest.toTransaction();
+
+    // should only includes MaxFee which is a required policy
+    expect(tx.policies.length).toBe(1);
+    expect(tx.policyTypes).toBe(PolicyType.MaxFee);
+
+    txRequest = new CreateTransactionRequest({ bytecodeWitnessIndex: 0 });
+
+    expect(txRequest.tip).toBeUndefined();
+    expect(txRequest.maturity).toBeUndefined();
+    expect(txRequest?.witnessLimit).toBeUndefined();
+
+    expect(txRequest.maxFee).toBeDefined();
+
+    tx = txRequest.toTransaction();
+
+    // should only includes MaxFee which is a required policy
+    expect(tx.policies.length).toBe(1);
+    expect(tx.policyTypes).toBe(PolicyType.MaxFee);
+  });
+
+  it('should ensure optional TX policies are not set with undesired values', () => {
+    let txRequest: BaseTransactionRequest = new ScriptTransactionRequest({ tip: 0, maturity: 0 });
+
+    txRequest = new ScriptTransactionRequest();
+
+    expect(txRequest.tip).toBeUndefined();
+    expect(txRequest.maturity).toBeUndefined();
+
+    expect(txRequest.maxFee).toBeDefined();
+
+    let tx = txRequest.toTransaction();
+
+    // should only includes maxFee which is a required policy
+    expect(tx.policies.length).toBe(1);
+    expect(tx.policyTypes).toBe(PolicyType.MaxFee);
+
+    txRequest = new CreateTransactionRequest({ tip: 0, maturity: 0, bytecodeWitnessIndex: 0 });
+
+    expect(txRequest.tip).toBeUndefined();
+    expect(txRequest.maturity).toBeUndefined();
+
+    expect(txRequest.maxFee).toBeDefined();
+
+    tx = txRequest.toTransaction();
+
+    // should only includes maxFee which is a required policy
+    expect(tx.policies.length).toBe(1);
+    expect(tx.policyTypes).toBe(PolicyType.MaxFee);
+  });
+
   it('should ensure TX policies are properly set (ScriptTransactionRequest)', async () => {
     const receiver = Wallet.generate({ provider });
 
@@ -65,7 +128,7 @@ describe('Policies', () => {
       tip: 10,
       maturity: randomNumber(1, 2),
       witnessLimit: randomNumber(800, 900),
-      maxFee: 200,
+      maxFee: 1000,
     };
 
     const txRequest = new ScriptTransactionRequest(txParams);
@@ -101,9 +164,9 @@ describe('Policies', () => {
 
     const txParams: CustomTxParams = {
       tip: 11,
-      witnessLimit: randomNumber(800, 900),
-      maturity: randomNumber(1, 2),
-      maxFee: 1500,
+      witnessLimit: 2000,
+      maturity: 1,
+      maxFee: 5000,
     };
 
     const { transactionRequest: txRequest } = factory.createTransactionRequest(txParams);
@@ -139,7 +202,7 @@ describe('Policies', () => {
       tip: 5,
       maturity: randomNumber(1, 2),
       witnessLimit: randomNumber(800, 900),
-      maxFee: 2500,
+      maxFee: 3000,
     });
 
     const txRequest = await callScope.getTransactionRequest();
@@ -169,7 +232,7 @@ describe('Policies', () => {
       tip: 2,
       maturity: randomNumber(1, 2),
       witnessLimit: randomNumber(800, 900),
-      maxFee: 2000,
+      maxFee: 3000,
     });
 
     const txRequest = await callScope.getTransactionRequest();
@@ -191,7 +254,7 @@ describe('Policies', () => {
       tip: 4,
       maturity: randomNumber(1, 2),
       witnessLimit: randomNumber(800, 900),
-      maxFee: 1800,
+      maxFee: 3000,
     };
 
     const pendingTx = await wallet.transfer(receiver.address, 500, baseAssetId, txParams);
@@ -219,7 +282,7 @@ describe('Policies', () => {
       tip: 1,
       maturity: randomNumber(1, 2),
       witnessLimit: randomNumber(800, 900),
-      maxFee: 2500,
+      maxFee: 3000,
     };
 
     const pendingTx = await wallet.transferToContract(contract.id, 500, baseAssetId, txParams);
@@ -237,7 +300,7 @@ describe('Policies', () => {
 
     const txParams: CustomTxParams = {
       maturity: randomNumber(1, 2),
-      witnessLimit: 5,
+      witnessLimit: 0,
     };
 
     await expect(async () => {