+++ packages/fuel-gauge/src/predicate/predicate-with-contract.test.ts
@@ -47,9 +47,9 @@ describe('Predicate', () => {
       });
       // Create a instance of the contract with the predicate as the caller Account
       const contractPredicate = new Contract(contract.id, contract.interface, predicate);
-      const predicateBalance = await fundPredicate(wallet, predicate, amountToPredicate);
+      await fundPredicate(wallet, predicate, amountToPredicate);
 
-      const { value } = await contractPredicate.functions
+      const { value, transactionResult } = await contractPredicate.functions
         .return_context_amount()
         .callParams({
           forward: [500, provider.getBaseAssetId()],
@@ -57,9 +57,7 @@ describe('Predicate', () => {
         .call();
 
       expect(value.toString()).toEqual('500');
-
-      const finalPredicateBalance = await predicate.getBalance();
-      expect(finalPredicateBalance.lt(predicateBalance)).toBeTruthy();
+      expect(transactionResult.isStatusSuccess).toBeTruthy();
     });
 
     it('calls a predicate and uses proceeds for a contract call', async () => {
@@ -73,7 +71,8 @@ describe('Predicate', () => {
       } = launcher;
       const receiver = WalletUnlocked.generate({ provider });
 
-      const initialReceiverBalance = toNumber(await receiver.getBalance());
+      const receiver = Wallet.generate({ provider });
+      const receiverInitialBalance = await receiver.getBalance();
 
       // calling the contract with the receiver account (no resources)
       contract.account = receiver;
@@ -95,11 +94,8 @@ describe('Predicate', () => {
           },
         ],
       });
-      const initialPredicateBalance = toNumber(await predicate.getBalance());
 
       await fundPredicate(wallet, predicate, amountToPredicate);
-
-      expect(toNumber(await predicate.getBalance())).toBeGreaterThan(initialPredicateBalance);
 
       // executing predicate to transfer resources to receiver
       const tx = await predicate.transfer(
@@ -107,35 +103,17 @@ describe('Predicate', () => {
         amountToReceiver,
         provider.getBaseAssetId()
       );
+      let { isStatusSuccess } = await tx.waitForResult();
+      expect(isStatusSuccess).toBeTruthy();
 
-      const { fee: predicateTxFee } = await tx.waitForResult();
+      const receiverFinalBalance = await receiver.getBalance();
+      expect(receiverFinalBalance.gt(receiverInitialBalance)).toBeTruthy();
 
-      // calling the contract with the receiver account (with resources)
-      const contractAmount = 10;
-      const {
-        transactionResult: { fee: receiverTxFee },
-      } = await contract.functions.mint_coins(200).call();
+      ({
+        transactionResult: { isStatusSuccess },
+      } = await contract.functions.mint_coins(200).call());
 
-      const finalReceiverBalance = toNumber(await receiver.getBalance());
-      const remainingPredicateBalance = toNumber(await predicate.getBalance());
-
-      const expectedFinalReceiverBalance =
-        initialReceiverBalance + amountToReceiver - contractAmount - receiverTxFee.toNumber();
-
-      expectToBeInRange({
-        value: finalReceiverBalance,
-        min: expectedFinalReceiverBalance - 20,
-        max: expectedFinalReceiverBalance + 20,
-      });
-
-      const expectedFinalPredicateBalance =
-        initialPredicateBalance + amountToPredicate - amountToReceiver - predicateTxFee.toNumber();
-
-      expectToBeInRange({
-        value: expectedFinalPredicateBalance,
-        min: remainingPredicateBalance - 20,
-        max: remainingPredicateBalance + 20,
-      });
+      expect(isStatusSuccess).toBeTruthy();
     });
   });
 });