+++ packages/crypto/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 0.84.0
+
+### Patch Changes
+
+- Updated dependencies [2990edb]
+  - @fuel-ts/utils@0.84.0
+  - @fuel-ts/errors@0.84.0
+
+## 0.83.0
+
+### Patch Changes
+
+- Updated dependencies [29f46ef]
+- Updated dependencies [9c3c094]
+- Updated dependencies [b026feb]
+  - @fuel-ts/utils@0.83.0
+  - @fuel-ts/errors@0.83.0
+
 ## 0.82.0
 
 ### Patch Changes