+++ packages/fuel-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @fuel-ts/fuel-core
 
+## 0.84.0
+
+### Minor Changes
+
+- 2990edb: chore!: upgrade `fuel-core` to `0.26.0`
+
+## 0.83.0
+
+### Minor Changes
+
+- 9c3c094: chore!: upgrade `fuel-core` to `0.24.3`
+
 ## 0.82.0
 
 ## 0.81.0