+++ packages/merkle/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 0.84.0
+
+### Patch Changes
+
+- @fuel-ts/hasher@0.84.0
+- @fuel-ts/math@0.84.0
+
+## 0.83.0
+
+### Patch Changes
+
+- @fuel-ts/hasher@0.83.0
+- @fuel-ts/math@0.83.0
+
 ## 0.82.0
 
 ### Patch Changes